copynumber: Segmentation of single- and multi-track copy number data by penalized least squares regression (with hg38 and mm10).

Penalized least squares regression is applied to fit piecewise constant curves to copy number data to locate genomic regions of constant copy number. Procedures are available for individual segmentation of each sample, joint segmentation of several samples and joint segmentation of the two data tracks from SNP-arrays. Several plotting functions are available for visualization of the data and the segmentation results.

See also

Author

Maintainer: Gro Nilsen gronilse@ifi.uio.no

Authors:

Gro Nilsen gronilse@ifi.uio.no

Knut Liestol

Ole Christian Lingjaerde

Igor Dolgalev igor.dolgalev@nyumc.org (ORCID)